Split widget: batch header row
