+++ extension/devtools/app/lib/main.dart
@@ -56,6 +56,7 @@ part 'widgets/computed_table_header.dart';
 part 'widgets/computed_row.dart';
 part 'widgets/computed_detail.dart';
 part 'widgets/batch_list.dart';
+part 'widgets/batch_header_row.dart';
 part 'widgets/panel_placeholder.dart';
 part 'widgets/panel_scroll_view.dart';
 part 'widgets/panel_state_cards.dart';
@@ -983,65 +984,6 @@ class _BatchingPanel extends StatelessWidget {
             );
           },
         ),
-      ),
-    );
-  }
-}
-
-class _BatchHeaderRow extends StatelessWidget {
-  const _BatchHeaderRow();
-
-  @override
-  Widget build(BuildContext context) {
-    final labelStyle = Theme.of(context).textTheme.labelSmall;
-
-    return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
-      decoration: BoxDecoration(
-        border: Border(
-          bottom: BorderSide(
-            color: Theme.of(context).dividerColor.withValues(alpha: 0.4),
-          ),
-        ),
-      ),
-      child: Row(
-        children: [
-          Expanded(
-            flex: 2,
-            child: Align(
-              alignment: Alignment.center,
-              child: Text('Batch', style: labelStyle),
-            ),
-          ),
-          Expanded(
-            flex: 2,
-            child: Align(
-              alignment: Alignment.center,
-              child: Text('Depth', style: labelStyle),
-            ),
-          ),
-          Expanded(
-            flex: 2,
-            child: Align(
-              alignment: Alignment.center,
-              child: Text('Writes', style: labelStyle),
-            ),
-          ),
-          Expanded(
-            flex: 2,
-            child: Align(
-              alignment: Alignment.center,
-              child: Text('Duration', style: labelStyle),
-            ),
-          ),
-          Expanded(
-            flex: 3,
-            child: Align(
-              alignment: Alignment.center,
-              child: Text('Ended', style: labelStyle),
-            ),
-          ),
-        ],
       ),
     );
   }
